+++ Sources/Descriptions/DIDescriptions.swift
@@ -14,54 +14,42 @@ extension DITypeInfo: CustomStringConvertible {
   }
 }
 
-extension DIResolveStyle: CustomStringConvertible {
-  public var description: String {
-    switch self {
-    case .one:
-      return "resolve one"
-    case .many:
-      return "resolve many"
-    case .byName(let name):
-      return "resolve by name: \(name)"
-    }
-  }
-}
-
 extension DIError: CustomStringConvertible {
   public var description: String {
-    switch self {
-    case .typeIsNotFound(let type):
-      return "Type: \(type) not found\n"
-    case .typeIsNotFoundForName(let type, let name, let typesInfo):
-      return "Cannot found type: \(type) for name: \(name).\n\tUse:\n\(multiLine(typesInfo))\n"
-    case .notSpecifiedInitializationMethodFor(let typeInfo):
-      return "Not specified initial method for:\n \(typeInfo)\n"
-    case .initializationMethodWithSignatureIsNotFoundFor(let typeInfo, let signature):
-      return "Initial method with signature: \(signature) not found.\n\tUse: \(typeInfo)\n"
-    case .pluralSpecifiedDefaultType(let type, let typesInfo):
-      return "Plural specified default type for type: \(type).\n\tUse:\n\(multiLine(typesInfo))\n"
-    case .defaultTypeIsNotSpecified(let type, let typesInfo):
-      return "Default type not specified for type: \(type).\n\tUse:\n\(multiLine(typesInfo))\n"
-    case .intersectionNamesForType(let type, let names, let typesInfo):
-      return "Intersection names for type: \(type).\n\tIntersections: \(names)\n\tUse:\n\(multiLine(typesInfo))\n"
-    case .typeIsIncorrect(let requestedType, let realType, let typeInfo):
-      return "Incorrect type.\n\tRequested type: \(requestedType)\n\tReal type: \(realType)\n\tUse: \(typeInfo)\n"
-    case .recursiveInitialization(let typeInfo):
-      return "Recursive initialization into:\(typeInfo)\n"
-    case .noAccess(let typesInfo, let modules):
+    #if ENABLE_DI_MODULE // into switch no works. or copy full switch...
+    if case .noAccess(let typesInfo, let modules) = self {
       return "No access to \(typesInfo). This type can resolve from: \(modules)\n"
+    }
+    #endif
+    
+    switch self {
+    /// Until Resolve
+    case .typeNotFound(let type):
+      return "Cannot found type: \(type)\n"
+    case .typeForNameNotFound(let type, let name, let typesInfo):
+      return "Cannot found type: \(type) for name: \(name).\n\tUse:\n\(multiLine(typesInfo))\n"
+    case .initialMethodNotFound(let typeInfo, let signature):
+      return "Cannot found initial method with signature: \(signature).\n\tUse: \(typeInfo)\n"
+    case .ambiguousType(let type, let typesInfo):
+      return "Ambiguous type: \(type).\n\tUse:\n\(multiLine(typesInfo))\n"
+    case .incorrectType(let requestedType, let realType, let typeInfo):
+      return "Incorrect type: \(realType) for requested type: \(requestedType)\n\tUse: \(typeInfo)\n"
+    case .recursiveInitial(let typeInfo):
+      return "Recursive initialize into type info: \(typeInfo)\n"
+      
+    /// Until Build
+    case .noSpecifiedInitialMethod(let typeInfo):
+      return "Not specified initial method for type info: \(typeInfo)\n"
+    case .intersectionNames(let type, let names, let typesInfo):
+      return "Intersection names for type: \(type).\n\tIntersections: \(names)\n\tUse:\n\(multiLine(typesInfo))\n"
+    case .pluralDefaultAd(let type, let typesInfo):
+      return "Plural default ad for type: \(type).\n\tUse:\n\(multiLine(typesInfo))\n"
+      
       
     case .build(let errors):
       return "\nList:\n\(multiLine(errors))\n"
-    case .stack(let type, let child, let resolveStyle):
-      if case .stack(_,_,_) = child {
-        return "\(child)\t\(type) use \(resolveStyle)\n"
-      }
-      return "\(child)Stack:\n\t\(type) use \(resolveStyle)\n"
-    case .byCall(let file, let line, let function, let stack):
-      return "\nBy call function: \(function) in file: \((file as NSString).lastPathComponent) on line: \(line).\nDescription: \(stack)"
-    case .whileCreateSingleton(let typeInfo, let stack):
-      return "\nWhile create singleton for \(typeInfo)\n\(stack)\n"
+    default:
+      fatalError("No initialize full error description switch")
     }
   }